requirements.txt
scikit-learn>=0.24.0
pandas>=1.2.0
Flask>=1.1.2
python-telegram-bot>=13.1